+++ packages/yak_test/test/testers/constructor_tester_test.dart
@@ -1,0 +1,7 @@
+import 'package:yak_test/yak_test.dart';
+
+class Foo {
+  const Foo();
+}
+
+void main() => constructorTester(Foo.new);