+++ packages/yak_utils/lib/src/all.dart
@@ -3,4 +3,3 @@ export 'extensions/all.dart';
 export 'functions/all.dart';
 export 'tools/all.dart';
 export 'types/all.dart';
-export 'mixin/all.dart';