+++ packages/yak_utils/lib/src/tools/union/union.dart
@@ -1,9 +1,8 @@
 import 'package:meta/meta.dart';
-import 'package:yak_utils/yak_utils.dart';
 
 /// a simple union class
 @immutable
-class Union<A, B> with JsonMixin {
+class Union<A, B> {
   final A left;
   final B right;
   const Union(this.left, this.right);
@@ -16,12 +15,4 @@ class Union<A, B> with JsonMixin {
   @override
   @nonVirtual
   int get hashCode => Object.hash(left, right, runtimeType);
-
-  @override
-  Json toJson() => {
-        '$runtimeType': {
-          'left': '$left',
-          'right': '$right',
-        }
-      };
 }