+++ packages/yak_utils/test/delegates/nullary_delegate_test.dart
@@ -1,24 +1,26 @@
+import 'package:mocktail/mocktail.dart';
 import 'package:test/test.dart';
+import 'package:yak_test/yak_test.dart';
 import 'package:yak_utils/yak_utils.dart';
 
-class _TestDelegate implements NullaryDelegate<int> {
-  const _TestDelegate();
-  @override
-  int call() => 0;
-}
+class _NullaryTester extends Mock implements NullaryDelegate<int> {}
+
+class _NullaryTesterAsync extends Mock implements NullaryDelegateAsync<int> {}
 
 void main() {
-  group('NullaryDelegate', () {
-    test(
-        'GIVEN NullaryDelegate '
-        'WHEN delegate.call '
-        'THEN returns a Nullary', () {
-      final delegate = _TestDelegate();
-      expect(
-        delegate.call,
-        isA<Nullary<int>>(),
-        reason: 'delegat.call should be predictable',
-      );
-    });
-  });
+  group(
+    'NullaryDelegate Test',
+    () {
+      constructorTester(_NullaryTester.new);
+      nullaryDelegateTest(_NullaryTester.new);
+    },
+  );
+
+  group(
+    'NullaryDelegateAsync Test',
+    () {
+      constructorTester(_NullaryTester.new);
+      nullaryDelegateAsyncTest(_NullaryTesterAsync.new);
+    },
+  );
 }