+++ packages/yak_tween/CHANGELOG.md
@@ -2,7 +2,7 @@
 - fix analysis
 - updated `example`
 - created from the ashes of `yak_math`
-- bump dependencies
+- update dependencies
 - updated readme