+++ packages/yak_utils/CHANGELOG.md
@@ -1,4 +1,8 @@
-### 1.0.0+14
+### 1.0.1+1
+- remove `JsonMixin`
+- update delegates tests
+
+### 1.0.0
 - update sets of `typedef` and `delegates`
 - add:
     - `MultiDelegate`
@@ -8,8 +12,7 @@
     - `"Void"` functions and delegates
     - plus multiple extensions
 - update readme
-- bump dependencies
-
+- update dependencies
 
 ### 0.1.0
 - updated `example` and `readme`